web/requirements.txt
# ==== Cloud Run Web App用 ====
google-cloud-storage
flask
gunicorn
python-dotenv
werkzeug
